check jwt token and get authorization data

+++ schema/src/api/chat-controller.ts
@@ -3,7 +3,8 @@
 // tslint:disable: array-type
 
 import express, { Router } from "express";
-import { Chat, ChatCreateData, Chats } from "./api-types";
+import { getAuthorization } from "../lib/api-controller-helpers";
+import { AuthorizationData, Chat, ChatCreateData, Chats } from "./api-types";
 
 export interface IChatService {
   listChats: () => Promise<Chats>;
@@ -12,11 +13,16 @@ export interface IChatService {
   deleteChat: (id: number) => Promise<void>;
 }
 
-export function createChatController(service: IChatService): Router {
+export function createChatController(service: IChatService, jwtSecret: string): Router {
   const router = express.Router();
 
   router.get("/chats", async (req, res) => {
     try {
+      const authorization = getAuthorization<AuthorizationData>(req, jwtSecret);
+      if (!authorization) {
+        res.status(403).send("forbidden");
+        return;
+      }
       const result = await service.listChats();
       res.json(result);
     } catch (e) {
@@ -27,6 +33,11 @@ export function createChatController(service: IChatService): Router {
 
   router.post("/chats", async (req, res) => {
     try {
+      const authorization = getAuthorization<AuthorizationData>(req, jwtSecret);
+      if (!authorization) {
+        res.status(403).send("forbidden");
+        return;
+      }
       const chatCreateData = req.body as ChatCreateData;
       const result = await service.createChat(chatCreateData);
       res.json(result);
@@ -38,6 +49,11 @@ export function createChatController(service: IChatService): Router {
 
   router.get("/chats/:id", async (req, res) => {
     try {
+      const authorization = getAuthorization<AuthorizationData>(req, jwtSecret);
+      if (!authorization) {
+        res.status(403).send("forbidden");
+        return;
+      }
       const id = parseInt(req.params.id as string, 10);
       const result = await service.getChat(id);
       res.json(result);
@@ -49,6 +65,11 @@ export function createChatController(service: IChatService): Router {
 
   router.delete("/chats/:id", async (req, res) => {
     try {
+      const authorization = getAuthorization<AuthorizationData>(req, jwtSecret);
+      if (!authorization) {
+        res.status(403).send("forbidden");
+        return;
+      }
       const id = parseInt(req.params.id as string, 10);
       await service.deleteChat(id);
     } catch (e) {
